Split the UI into a separate repository.
