--- .babelrc
@@ -1,5 +1,0 @@
-{
-	"presets": [
-		"solid"
-	]
-}
