+++ .prettierrc.yaml
@@ -2,6 +2,3 @@
 
 # Don't insert semi-colons unless needed
 semi: false
-
-plugins:
-  - "prettier-plugin-tailwindcss"
