--- src/common/tsconfig.json
@@ -1,12 +1,0 @@
-{
-	"extends": "../tsconfig.json",
-	"include": [
-		"."
-	],
-	"compilerOptions": {
-		"types": [
-			"mp4box",
-			"web"
-		],
-	},
-}
