--- src/contribute/tsconfig.json
@@ -1,21 +1,0 @@
-{
-	"extends": "../tsconfig.json",
-	"include": [
-		"."
-	],
-	"compilerOptions": {
-		"types": [
-			"dom-mediacapture-transform",
-			"dom-webcodecs",
-			"web"
-		]
-	},
-	"references": [
-		{
-			"path": "../common"
-		},
-		{
-			"path": "../transport"
-		},
-	]
-}
