+++ src/playback/port.ts
@@ -1,4 +1,4 @@
-import * as Message from "./message"
+import * as Message from "./worker/message"
 
 export type Callback = (e: Message.FromWorker) => void
 
@@ -10,7 +10,7 @@ export class Port {
 	#callback: Callback
 
 	constructor(callback: Callback) {
-		const url = new URL("worker.ts", import.meta.url)
+		const url = new URL("worker/index.ts", import.meta.url)
 
 		this.#callback = callback
