+++ src/playback/worker/container.ts
@@ -1,4 +1,4 @@
-import * as MP4 from "../common/mp4"
+import * as MP4 from "../../common/mp4"
 import { Frame } from "./timeline"
 
 // Decode a MP4 container into individual frames.
