+++ src/transport/client.ts
@@ -4,7 +4,7 @@ import * as Control from "./control"
 import { Objects } from "./object"
 import { Connection } from "./connection"
 
-export interface Config {
+export interface ClientConfig {
 	url: string
 
 	// Parameters used to create the MoQ session
@@ -15,54 +15,63 @@ export interface Config {
 	fingerprint?: string
 }
 
-export async function connect(config: Config) {
-	// Helper function to make creating a promise easier
-	const options: WebTransportOptions = {}
+export class Client {
+	#config: ClientConfig
+	#fingerprint: Promise<WebTransportHash | undefined>
 
-	if (config.fingerprint) {
-		try {
-			const fingerprint = await fetchFingerprint(config.fingerprint)
-			options.serverCertificateHashes = [fingerprint]
-		} catch (e) {
+	constructor(config: ClientConfig) {
+		this.#config = config
+		this.#fingerprint = this.#fetchFingerprint(config.fingerprint).catch((e) => {
 			console.warn("failed to fetch fingerprint: ", e)
+			return undefined
+		})
+	}
+
+	async connect(): Promise<Connection> {
+		// Helper function to make creating a promise easier
+		const options: WebTransportOptions = {}
+
+		const fingerprint = await this.#fingerprint
+		if (fingerprint) options.serverCertificateHashes = [fingerprint]
+
+		const quic = new WebTransport(this.#config.url, options)
+		await quic.ready
+
+		const stream = await quic.createBidirectionalStream()
+
+		const writer = new Stream.Writer(stream.writable)
+		const reader = new Stream.Reader(stream.readable)
+
+		const setup = new Setup.Stream(reader, writer)
+
+		// Send the setup message.
+		await setup.send.client({ versions: [Setup.Version.DRAFT_00], role: this.#config.role })
+
+		// Receive the setup message.
+		// TODO verify the SETUP response.
+		const _server = await setup.recv.server()
+
+		const control = new Control.Stream(reader, writer)
+		const objects = new Objects(quic)
+
+		return new Connection(quic, control, objects)
+	}
+
+	async #fetchFingerprint(url?: string): Promise<WebTransportHash | undefined> {
+		if (!url) return
+
+		// TODO remove this fingerprint when Chrome WebTransport accepts the system CA
+		const response = await fetch(url)
+		const hexString = await response.text()
+
+		const hexBytes = new Uint8Array(hexString.length / 2)
+		for (let i = 0; i < hexBytes.length; i += 1) {
+			hexBytes[i] = parseInt(hexString.slice(2 * i, 2 * i + 2), 16)
 		}
-	}
 
-	const quic = new WebTransport(config.url, options)
-	await quic.ready
-
-	const stream = await quic.createBidirectionalStream()
-
-	const writer = new Stream.Writer(stream.writable)
-	const reader = new Stream.Reader(stream.readable)
-
-	const setup = new Setup.Stream(reader, writer)
-
-	// Send the setup message.
-	await setup.send.client({ versions: [Setup.Version.DRAFT_00], role: config.role })
-
-	// Receive the setup message.
-	// TODO verify the SETUP response.
-	const _server = await setup.recv.server()
-
-	const control = new Control.Stream(reader, writer)
-	const objects = new Objects(quic)
-
-	return new Connection(quic, control, objects)
-}
-
-async function fetchFingerprint(url: string): Promise<WebTransportHash> {
-	// TODO remove this fingerprint when Chrome WebTransport accepts the system CA
-	const response = await fetch(url)
-	const hexString = await response.text()
-
-	const hexBytes = new Uint8Array(hexString.length / 2)
-	for (let i = 0; i < hexBytes.length; i += 1) {
-		hexBytes[i] = parseInt(hexString.slice(2 * i, 2 * i + 2), 16)
-	}
-
-	return {
-		algorithm: "sha-256",
-		value: hexBytes,
+		return {
+			algorithm: "sha-256",
+			value: hexBytes,
+		}
 	}
 }
