--- src/transport/tsconfig.json
@@ -1,16 +1,0 @@
-{
-	"extends": "../tsconfig.json",
-	"include": [
-		".",
-	],
-	"compilerOptions": {
-		"types": [
-			"web"
-		]
-	},
-	"references": [
-		{
-			"path": "../common"
-		},
-	],
-}
