--- src/ui/tsconfig.json
@@ -1,23 +1,0 @@
-{
-	"extends": "../tsconfig.json",
-	"include": [
-		"."
-	],
-	"compilerOptions": {
-		"types": [
-			"mp4box",
-			"web",
-			"parcel-bundler"
-		],
-		"jsx": "preserve",
-		"jsxImportSource": "solid-js",
-	},
-	"references": [
-		{
-			"path": "../playback"
-		},
-		{
-			"path": "../contribute"
-		},
-	],
-}
